refactor(initramfs): Clean up network code
