+++ internal/app/init/pkg/network/constants.go
@@ -1,0 +1,10 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+package network
+
+// DefaultInterface is the default public interface for
+// the host that will be used for dhcp requests
+// TODO replace this with kernel arg
+const DefaultInterface = "eth0"
